+++ app/jobs/rails_audit_log/streaming/publish_entry_job.rb
@@ -1,0 +1,13 @@
+module RailsAuditLog
+  module Streaming
+    class PublishEntryJob < ApplicationJob
+      def perform(entry_attrs)
+        entry = AuditLogEntry.new(entry_attrs)
+        ActiveSupport::Notifications.instrument(
+          NotificationsAdapter::EVENT,
+          entry: entry
+        )
+      end
+    end
+  end
+end